+++ src/OpenTelemetry.Instrumentation.AWS/CHANGELOG.md
@@ -8,7 +8,7 @@
 * Add the `aws.s3.bucket` and `aws.s3.key` attributes to S3 spans.
   ([#4029](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4029))
 
-* Add instrumentation scope version and schema URL to traces.
+* Add instrumentation scope version and schema URL to metrics and traces.
   ([#4063](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4063))
 
 ## 1.15.0